+++ SYU/src/main/java/com/ddo/java/Start.java
@@ -1,2 +1,25 @@
-package com.ddo.java;public class Start {
+package com.ddo.java;
+
+import java.util.Scanner;
+
+public class Start {
+    public void start (int hp, Student player){
+        System.out.print("닉네임을 입력해주세요: ");
+        Scanner sc = new Scanner(System.in);
+        String name = sc.next();
+
+        // 캐릭터 생성
+        for (int day = 1; day <= 14; day++) { // 14일간 게임 진행
+            System.out.println("Day " + day); // 날짜 출력
+
+            // 캐릭터 상태 출력
+            if (day % 4 == 0) {
+                player.displayStats();
+            }
+
+            // 아침 이벤트 실행
+            Morning morning = new Morning();
+            morning.runMonrningEvent(player);
+        }
+    }
 }
