--- moring-wakeup-and-class/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Thu Mar 21 23:16:05 KST 2024
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.4-bin.zip
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
