--- moring-wakeup-and-class/settings.gradle
@@ -1,2 +1,0 @@
-rootProject.name = 'moring-wakeup-and-class'
-
